# [DataFrame] Speed up dtypes