## Summary

Speeds up dtypes. In order to save memory, this PR gets smaller dataframes from the partitions which are merged together to find `dtypes`. This saves memory and offers performance benefits because it does not create a new version of `_block_partitions` with similar data.

This PR deprecates #2088.

## Performance tests on 762 MB of data
### Code (copied from iPython)
```python
frame_data = np.random.randint(0,100,size=(10**6, 100))

%%time
df = pd.DataFrame(frame_data)
df.dropna(inplace=True)
repr(df)
```

### On Master
CPU times: user 850 ms, sys: 1.1 s, total: 1.95 s
Wall time: 3.35 s

Number of objects in object table = 336
Total size of objects in object table in MB: 5456.237922668457

### Master modified so dtypes doesn't access Ray/the object store
CPU times: user 851 ms, sys: 1.08 s, total: 1.93 s
Wall time: 3.01 s

Number of objects in object table = 256
Total size of objects in object table in MB: 4631.916297912598

### Current PR
CPU times: user 910 ms, sys: 1.08 s, total: 1.99 s
Wall time: 3.13 s

Number of objects in object table = 393
Total size of objects in object table in MB: 4632.369149208069