@pschafhalter 2018-05-22
Redesigned speedup according to discussion with @kunalgosar. Also re-ran performance tests on a wider dataframe:

### Code
```python
frame_data = np.random.randint(0,100,size=(10000, 10000))

%%time
df = pd.DataFrame(frame_data)
df.dropna(inplace=True)
repr(df)
```

### Master

CPU times: user 500 ms, sys: 826 ms, total: 1.33 s
Wall time: 2.14 s

Number of objects in object table = 336
Total size of objects in object table in MB: 5346.965839385986

### Master modified so dtypes doesn't access Ray/the object store
Note: this returns an incorrect result for dtypes and should be the lower bound for this PR.

CPU times: user 433 ms, sys: 890 ms, total: 1.32 s
Wall time: 1.81 s

Number of objects in object table = 256
Total size of objects in object table in MB: 4582.445110321045

### Previous commit

CPU times: user 562 ms, sys: 804 ms, total: 1.37 s
Wall time: 1.92 s

Number of objects in object table = 393
Total size of objects in object table in MB: 4584.221190452576

### Current commit

CPU times: user 464 ms, sys: 846 ms, total: 1.31 s
Wall time: 1.82 s

Number of objects in object table = 264
Total size of objects in object table in MB: 4582.500843048096