@pschafhalter 2018-05-23
All tests passed on private travis.